Fix provider health check timeout preventing status patch Our confighealth reconciler performs a health check HTTP request to the anynines service broker and writes the result to the providerconfig's status field.
Previously, one Golang context with a timeout was used for both the health check and the subsequent patching of providerconfig's status. However, if the health check request exceeded the timeout, the patching of the status field could never succeed. client-go's Patch() method would error with a "context deadline exceeded" error message.

This commit introduces a separate sub-context for the health check request, which has an explicit timeout, while using the parent context (which does not time out) for patching the status field.